+++ src/features/CardList/CardListElements/CardList.jsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { cn } from "../../utils";
-import BG_COLOR from "../../constants/backgroundcolor";
-import BG_PATTERN from "../../constants/backgroundpattern";
+import { cn } from "../../../utils";
+import BG_COLOR from "../../../constants/backgroundcolor";
+import BG_PATTERN from "../../../constants/backgroundpattern";
 
 const CardList = ({ item }) => {
   return (
