+++ src/features/CardList/CardListSet.jsx
@@ -1,4 +1,4 @@
-import CardList from "./CardList";
+import CardList from "./CardListElements/CardList";
 import react, { useEffect, useState } from "react";
 import PureList from "./CardListElements/PureList";
 import { cn } from "../../utils";
